+++ src/ui/svelte/weg/components/Dock.svelte
@@ -1,25 +1,24 @@
 <script lang="ts">
   import { invoke, SeelenCommand } from "@seelen-ui/lib";
-  import { SeelenWegMode, WegItemType, WegPinnedItemsVisibility, WegTemporalItemsVisibility } from "@seelen-ui/lib/types";
+  import {
+    SeelenWegMode,
+    WegItemType,
+    WegPinnedItemsVisibility,
+    WegTemporalItemsVisibility,
+  } from "@seelen-ui/lib/types";
   import { DragDropProvider, DragOverlay } from "@dnd-kit/svelte";
   import { move } from "@dnd-kit/helpers";
   import { BackgroundByLayers } from "libs/ui/svelte/components/BackgroundByLayers";
   import { t } from "../i18n/index.ts";
   import { dockState } from "../state/items.svelte.ts";
-  import { settingsState, getDockContextMenuAlignment } from "../state/settings.svelte.ts";
+  import { settingsState } from "../state/settings.svelte.ts";
   import { systemState } from "../state/system.svelte.ts";
   import { interactables, getWindowsForItem } from "../state/windows.svelte.ts";
   import { dockShouldBeHidden, setDockIsDraggingItem } from "../state/hidden.svelte.ts";
   import { getSeelenWegMenu } from "../dockMenu.ts";
   import { DND_PLUGINS, DND_SENSORS } from "libs/ui/dnd.ts";
-  import type { SwItem } from "../types.ts";
   import DraggableItem from "./DraggableItem.svelte";
-  import Separator from "./items/Separator.svelte";
-  import StartMenu from "./items/StartMenu.svelte";
-  import ShowDesktop from "./items/ShowDesktop.svelte";
-  import RecycleBin from "./items/RecycleBin.svelte";
-  import MediaSession from "./items/MediaSession.svelte";
-  import UserApplication from "./items/UserApplication.svelte";
+  import WegItemSwitch from "./WegItemSwitch.svelte";
 
   const settings = $derived(settingsState.value as any);
   const isHorizontal = $derived(
@@ -56,7 +55,8 @@
   );
 
   function onContextMenu() {
-    const { alignX, alignY } = getDockContextMenuAlignment(settingsState.position);
+    const alignX = settingsState.popupAlignX;
+    const alignY = settingsState.popupAlignY;
     invoke(SeelenCommand.TriggerContextMenu, {
       menu: { ...getSeelenWegMenu($t), alignX, alignY },
       forwardTo: null,
@@ -103,19 +103,7 @@
         {:else}
           {#each visibleItems as item, index (item.id)}
             <DraggableItem {item} {index}>
-              {#if item.type === WegItemType.AppOrFile}
-                <UserApplication {item} />
-              {:else if item.type === WegItemType.StartMenu}
-                <StartMenu {item} />
-              {:else if item.type === WegItemType.ShowDesktop}
-                <ShowDesktop {item} />
-              {:else if item.type === WegItemType.Media}
-                <MediaSession {item} />
-              {:else if item.type === WegItemType.Separator}
-                <Separator {item} />
-              {:else if item.type === WegItemType.TrashBin}
-                <RecycleBin {item} />
-              {/if}
+              <WegItemSwitch {item} />
             </DraggableItem>
           {/each}
         {/if}
@@ -125,19 +113,7 @@
         {#snippet children(source)}
           {@const overlayItem = visibleItems.find((c) => c.id === source.id)}
           {#if overlayItem}
-            {#if overlayItem.type === WegItemType.AppOrFile}
-              <UserApplication item={overlayItem} isOverlay={true} />
-            {:else if overlayItem.type === WegItemType.StartMenu}
-              <StartMenu item={overlayItem} />
-            {:else if overlayItem.type === WegItemType.ShowDesktop}
-              <ShowDesktop item={overlayItem} />
-            {:else if overlayItem.type === WegItemType.Media}
-              <MediaSession item={overlayItem} />
-            {:else if overlayItem.type === WegItemType.Separator}
-              <Separator item={overlayItem} />
-            {:else if overlayItem.type === WegItemType.TrashBin}
-              <RecycleBin item={overlayItem} />
-            {/if}
+            <WegItemSwitch item={overlayItem} isOverlay={true} />
           {/if}
         {/snippet}
       </DragOverlay>